Separate mapping code per system components
