+++ packages/arb-bridge-eth/src/interface/build.json
@@ -1,5 +1,5 @@
 {
-  "rollupPreprocessor": "  - kind: ethereum/contract\n    name: IRollupCore\n    network: {{ l1Network }}\n    source:\n      address: '{{ rollupAddress }}'\n      abi: IRollupCore\n      startBlock: {{ rollupDeploymentBlock }}\n    mapping:\n      kind: ethereum/events\n      apiVersion: 0.0.5\n      language: wasm/assemblyscript\n      entities:\n        - Node\n      abis:\n        - name: IRollupCore\n          file: ./abis/IRollupCore.json\n      eventHandlers:\n        - event: NodeCreated(indexed uint256,indexed bytes32,bytes32,bytes32,uint256,uint256,bytes32,bytes32[3][2],uint256[4][2])\n          handler: handleNodeCreated\n        - event: NodeConfirmed(indexed uint256,bytes32,uint256,bytes32,uint256)\n          handler: handleNodeConfirmed\n        - event: NodeRejected(indexed uint256)\n          handler: handleNodeRejected\n      file: ./src/mapping.ts",
+  "rollupPreprocessor": "  - kind: ethereum/contract\n    name: IRollupCore\n    network: {{ l1Network }}\n    source:\n      address: '{{ rollupAddress }}'\n      abi: IRollupCore\n      startBlock: {{ rollupDeploymentBlock }}\n    mapping:\n      kind: ethereum/events\n      apiVersion: 0.0.5\n      language: wasm/assemblyscript\n      entities:\n        - Node\n      abis:\n        - name: IRollupCore\n          file: ./abis/IRollupCore.json\n      eventHandlers:\n        - event: NodeCreated(indexed uint256,indexed bytes32,bytes32,bytes32,uint256,uint256,bytes32,bytes32[3][2],uint256[4][2])\n          handler: handleNodeCreated\n        - event: NodeConfirmed(indexed uint256,bytes32,uint256,bytes32,uint256)\n          handler: handleNodeConfirmed\n        - event: NodeRejected(indexed uint256)\n          handler: handleNodeRejected\n      file: ./src/rollup.ts",
 
   "l1Network": "{{ l1Network }}",
   "outboxDeploymentBlock": "{{ outboxDeploymentBlock }}",
